requirements.txt
asgiref==3.2.7
Django==3.0.5
django-crispy-forms==1.9.0
Pillow==6.2.2
psycopg2==2.8.5
pytz==2019.3
sqlparse==0.3.1
